Move and rename class which is used as the response content type
